requirements.txt
flask
langchain
requests
Pillow
langchain-community
langchain-google-genai
opencv-python
google-generativeai
